+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v4.8.0
 
 - Supported `POST account/update_profile_image`. ([#617](https://github.com/twitter-dart/twitter-api-v2/issues/617))
+- Supported `POST account/update_profile_banner`. ([#614](https://github.com/twitter-dart/twitter-api-v2/issues/614))
 
 ## v4.7.1